Cleanup code from review.
