+++ frontend/src/interfaces/IMetaMaskError.ts
@@ -1,0 +1,4 @@
+export interface IMetaMaskError {
+  code: number;
+  message: string;
+}
