+++ frontend/src/interfaces/IWalletConnection.ts
@@ -1,0 +1,11 @@
+export interface IWalletConnection {
+  connected: boolean;
+  account: string;
+  provider: string;
+}
+
+export interface IWalletConnectionContext {
+  walletConnection: IWalletConnection;
+  setWalletConnection: (arg0: IWalletConnection) => {};
+}
+
