--- frontend/src/lib/wallet/WalletConnection.ts
@@ -1,21 +1,0 @@
-export interface WalletConnection {
-  connected: boolean;
-  account: string;
-  provider: string;
-}
-
-export interface WalletConnectionContext {
-  walletConnection: WalletConnection;
-  setWalletConnection: (arg0: WalletConnection) => {};
-}
-
-export const DefaultWalletConnection = {
-  connected: false,
-  account: '',
-  provider: '',
-};
-
-export const DefaultWalletConnectionContext = {
-  walletConnection: DefaultWalletConnection,
-  setWalletConnection: (arg0: WalletConnection) => {},
-};
